+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/request/ImageOptions.common.kt
@@ -129,7 +129,7 @@ data class ImageOptions(
     /**
      * Disk caching policy for Bitmaps affected by [sizeResolver] or [transformations]
      *
-     * @see com.github.panpf.sketch.cache.internal.ResultCacheDecodeInterceptor
+     * @see com.github.panpf.sketch.cache.internal.ResultCacheRequestInterceptor
      */
     val resultCachePolicy: CachePolicy?,