+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
   stream. [#272](https://github.com/panpf/sketch/issues/272)
 * fix: Fixed a bug where non-Android platforms used Skia to decode GIFs when the frame duration was
   0 and not modified to the default value of 100. [#273](https://github.com/panpf/sketch/issues/273)
+* change: ResultCacheDecodeInterceptor change to ResultCacheRequestInterceptor
 
 # 4.4.0-alpha01